fix: 將驗證的回應改成和api文件相符,將註冊email驗證邏輯放到register
