+++ app/Http/Controllers/RegisterController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Auth\Events\Registered;
+use Illuminate\Auth\Events\Verified;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 
@@ -67,4 +69,54 @@ class RegisterController extends Controller
 
         return response(['message' => 'User registered successfully!', 'user' => $user], 201);
     }
+
+
+
+        /**
+     * 註冊email驗證信確認
+     * 
+    
+     * 電子郵件驗證確認
+     *
+     * 此端點用於確認用戶的電子郵件驗證(和前端較無關聯）。
+     * 它會比對提供的hash值和用戶的電子郵件生成的hash值。
+     * 如果驗證成功，該用戶的電子郵件將被標記為已驗證，並且將觸發一個已驗證的事件。
+     * 系統會將email驗證的日期存入資料庫
+     *
+     * @param Request $request HTTP請求
+     * @param int $id 用戶ID
+     * @param string $hash 從驗證郵件中提供的hash值
+     * 
+     * @throws AuthorizationException 當提供的hash值不匹配時
+     * 
+     * @response 200 {
+     *   "message": "Email verified successfully."
+     * }
+     * 
+     * @response 200 {
+     *   "message": "Email already verified."
+     * }
+     */
+
+     public function verifyEmail(Request $request, $id, $hash)
+     {
+         $user = User::findOrFail($id);
+ 
+         // 此方法通常用來判斷文件是否被串改
+         if (!hash_equals((string) $hash, sha1($user->getEmailForVerification()))) {
+             throw new AuthorizationException();
+         }
+ 
+         // 判斷這個email是否已經驗證過
+         if ($user->hasVerifiedEmail()) {
+             return response(['message' => 'Email already verified.']);
+         }
+ 
+         // 到這一步就去將他的email驗證
+         if ($user->markEmailAsVerified()) {
+             event(new Verified($user));
+         }
+ 
+         return response(['message' => 'Email verified successfully.']);
+     }
 }
